Made exp__ calculate epsilon-propagation in the log domain.
I.e., x- rather than y-domain; for the other continuous primitives this was the default already anyway.

This prevents floating-point errors: particularly for negative arguments, the function converges obviously and quickly appraches const 0, yet the underflow would in the previous implementation sometimes cause reasonable result margins be mapped to 0 instead of |x|.

For positive arguments, the contrary was enforced.