Allowed for parametric plots as a simple special case of plots to a nontrivial vector space (ℝ²).
This is at the moment only usable by explicitly linear-combining 'constCntnFuncValue' basis vectors, which is cumbersome but works.

Also, the 'finiteGraphContinℝtoℝ²' function for visualising these plots is not quite done: it guarantees to properly render a single connected component of the curve, but fails when multiple (or none) such components are in view.