Toyed with a functor instance of `Region`.
Well, actually just with a function of `fmap`'s signature in `Functor (Region s) (Differentiable s) (->)`. That actual instance doesn't work already because of an overlap with the generic `Hask` instance of the constrained-categories `Functor` class; but anyways the implementation of that function is mathematically less straightforward than you might think, so don't continue this route for the time being.