We don't really need (A'A)⁻¹A' left-inverse to generate the projection to a Hilbert embedding.
In fact it's not really good, because it merely minimises the distance-squared to the embedded space WRT to the standard euclidean norm, which isn't generally meaningful for our applications.

A better alternative is to just use _pair_ of vectors and covectors, which make up an "orthonormal" basis in the obvious way, and make them simply row resp. column vectors of the transformation matrices. This is the obvious generalisation to what you'd do in the Euclidean case, doesn't require inverses or similar, and makes only reference to the norm induced by the given metric, not any inner product given (or assumed) by the space itself.