Get rid of `|-~.`, `.+~|` and `betweenBounds`.
These were supposed to deal crudely with boundary points, but we hadn't used them yet and – let's face it – weren't going to, as these functions are just too crude and wouldn't have allowed useful interpolation etc. anyway.

In the future, the `Geodesic` and `Riemannian` classes are supposed to fulfill this need.