Simplify `Affine` implementation.
The previous one went to great lengths to ensure that we don't depend on
a choice of reference point, but for a category instance this turned
out just impractical. Now implicitly depend on a reference point
via `Atlas`.

This also gives a somewhat nice generalisation of affine functions
on non-connected spaces ("layer vector spaces").