Local quadratic models at last work again, without extrem-deviation proxies.
Formerly, these quadratic models would not be calculated from the apriori information
per se but from extreme values of their uncertainty range. This makes the entire
regression highly expensive.

Now we only use extreme-proxies in calculation of the model uncertainty. This actually
turned out to be tricky, if we don't want to incur a great new workload.