Non-heuristic way of distinguishing web-boundary elements.
This one explicitly tries to construct (using a Givens-Rotation–like technique)
a half-space that contains all neighbour elements. If that succeeds, it
follows that the node borders a half space that lies (at least locally)
outside of the web.

This technique yields much fewer false-positive boundary nodes
for the random-scatter test cases, and perfectly describes the
boundary of the hexagonal test. It does however show up some
problems in the rectangular case.

In the future, we might base the web construction itself on this technique,
which seems to be rather more reliable than what we have now.