Choose a more purely-functional layout for `PointsWeb`.
The previous implementation with its flat-array storage was never too elegant,
and also turns out to be not good perfomance-wise with algorithms using `TreewiseTraverse`.

Now switch to a completely tree-based representation, with only _relative offset indices_
as network links.

Many of the functions aren't implemented yet; we want to do this in a new way anyway.