Native position-associated-value support in `ShadeTree`.
Previously, we used the `WithAny` hack to contain key-value pairs in the
off-the-bat key-only `ShadeTree` type. This works as a hack, but is not very
satisfying, especially now we want to use it as an associative container
also within `PointsWeb` (which so far has kept the associated values separately,
but we now want to change that).