Preclude differential equations from directly depending on 2nd derivatives.
While this is necessary for solving parabolic PDEs, it stands to reason that these
simply /cannot/ be solved in the way we would like (because of nonlocality), so
it's probably more prudent to restrict ourselves to hyperbolic / elliptic PDEs.

By not depending on the 2nd derivatives, we can again simplify the locally
fitted model to an affine one.