Manifolds with boundary as an independent class.
The previous hierarchy, in which manifolds without boundary were seen as a special case
of manifolds that are equal to their interior, introduced a lot of unnecessary complexity.

Now a manifold-with-boundary is just not a manifold at all itself, but merely seperates in
two spaces (interior and boundary) which both are manifolds.